Added file creation within root external storage
Instead of saving within the app’s external dir (i.e. within
emulated/sdCard0/Android/com…package/), we save to
Environment.getExternalStorage() allowing ease of access. This way, we
created a directory ‘CriminalIntent’ in the root of our device’s
external storage directory, in this case the path is as follows:
“emulated/0/CriminalIntent/crimes.json”

Note 1: we have to called MediaScannerConnection.scan() to make the
file visible in the files directory
Note 2: permissions are needed in the Android Manifest to write to
external storage
Note 3: we need to create the directory using File.mkdirs()