[Ch20 - update] Wired up ImageView to display photo
Added:
+ helper method cleanImageView() to free any memory used by Bitmap (or
at least, ready it for Garbage Collection)
+ wired up ImageView to display image associated with Crime

Note:
+ For sake of memory, clean up Bitmap after usage
+ Bitmap.recycle() frees the native storage of our bitmap
+ Rule of thumb, if images are large call recycle()
+ We loaded and unloaded the images in onStart() and onStop()
respectively. But why not onResume() and onPause()? This is because a
paused activity can still be partly visible i.e. when there is an
activity on top of it not covering the entire screen. onStop() and
onStart() lifecycle callbacks directly imply the the user is no longer
seeing the activity